fix(LottoPrize): calculatePrize long 타입을 받도록 수정

+++ java-lotto/src/main/java/lotto/domain/LottoPrize.java
@@ -48,7 +48,7 @@ public enum LottoPrize {
         return lottoPrize;
     }
 
-    public long calculatePrize(int count) {
+    public long calculatePrize(long count) {
         return this.prize * count;
     }
